Simplify the reverse transport: 8 relay files/3 tests -> 5/1, drop fingerprinting
The mechanism was over-built. Collapsing it, with the same real files-pull still
green end to end over a JSON wire:

- Drop request/response fingerprinting (command_id/md5) and the build_url
  cache-buster special-case. The exchange is strictly sequential and the importer
  resumes deterministically, so "the delivered result answers the first request
  on re-entry" holds without matching — a single consumed flag is enough.
- Merge the three remote classes (RelayImportDriver + RelayExchange +
  RelayExchangeEndpoint) into RelayExchange::handle_json: decode JSON, re-enter
  the real importer until it yields or completes, encode JSON.
- Merge the two source classes (RelaySourceWorker + RelayExportSource) into
  RelaySource: the outbound loop plus command->request translation and gunzip.
- Keep RelayTransport: it holds the one delivered result and a consumed flag that
  must survive the fresh-client re-entry loop (moving it onto the recreated
  client would let the second pass re-consume the result at an advanced cursor).
  This is the one abstraction that earns its keep; documented as such.
- Collapse the demo + array-level pull + wire tests into one real files-pull over
  the JSON boundary (ReverseTransportTest).

Relay lib 474 -> 262 lines (8 -> 5 files incl. load), relay tests 644 -> 174
lines (3 -> 1 test). import.php back to 504/74 PHPCS, direct mode untouched.
