+++ packages/reprint-importer/src/lib/relay/load.php
@@ -2,17 +2,16 @@
 /**
  * Loader for the reverse-transport relay classes.
  *
- * These implement "push as a remote-driven pull": the remote runs the importer
- * and owns all state; the local source only makes outbound requests. See
+ * "Push as a remote-driven pull": the remote runs the importer and owns all
+ * state; the local source only makes outbound requests. See
  * markdown/REVERSE-TRANSPORT.md.
  *
- * Classes are loaded leaf-first.
+ * RelayTransport (with TransportYield) is the seam the importer talks to;
+ * RelayExchange is the remote endpoint; RelaySource is the outbound-only local
+ * side (standalone — it needs neither the importer nor the exchange).
  */
 
 require_once __DIR__ . '/class-transport-yield.php';
 require_once __DIR__ . '/class-relay-transport.php';
 require_once __DIR__ . '/class-relay-exchange.php';
-require_once __DIR__ . '/class-relay-source-worker.php';
-require_once __DIR__ . '/class-relay-export-source.php';
-require_once __DIR__ . '/class-relay-import-driver.php';
-require_once __DIR__ . '/class-relay-exchange-endpoint.php';
+require_once __DIR__ . '/class-relay-source.php';
